+++ ocaml-lsp-server/src/dune_pull_diagnostics.mli
@@ -1,0 +1,8 @@
+open Import
+
+val identifier : string
+
+val document
+  :  Diagnostic.t list
+  -> DocumentDiagnosticParams.t
+  -> DocumentDiagnosticReport.t